bug!: throw an exception on invalid or duplicate headers
Previously, it was possible to specify the same header twice, and the "winner" would be non-deterministic since maps are inherently unordered. This updates the parsing to throw an error if it sees the same header key twice.

Additionally, parsing will now throw an error when given a custom metadata value that is not prefixed with "x-goog-meta-" (this was previously a warning).

Finally, this fixes a bug where header values with multiple consecutive colon characters (":") would be squashed into a single character during parsing.