chore: switch from globby to tiny-glob
tiny-glob has no fewer deps and much much faster. More importantly, it does NOT require es6 modules like globby and fast-glob do, which isn't currently compatible with how we build our Actions. We've been floating an old version of globby for awhile due to the latter, and it would be good to get on a dependency that is both lighter and compatible with our process.