feat: refactor uploading and ignore processing
This refactors the implementation for listing, ignoring, and uploading files.

The previous implementation relied on the UploadHelper to process files and globs, and it led to a lot of translation between relative and absolute paths. It also added a lot of layers of indirection and a really tight coupling between the filesystem and the uploader.

The new implementation pre-processes all files and ignores first, then delegates uploading on the complete list of files. This means the uploader does not need to behave differently when given a file versus a directory. This makes it easier to stub only the calls to storage.bucket.upload and truly exercise the entire pipeline in unit tests.

As part of this refactor, I dropped the pMap dependency. We've been pinned at an old version of this dependency because the new version is incompatible. There's a new function, inParallel that executes Promises in parallel. I'll move it to actions-utils after more battle testing.